+++ packages/ai/src/models.generated.ts
@@ -8269,8 +8269,8 @@ export const MODELS = {
 			reasoning: true,
 			input: ["text"],
 			cost: {
-				input: 0.21,
-				output: 0.7899999999999999,
+				input: 0.27,
+				output: 0.95,
 				cacheRead: 0.13,
 				cacheWrite: 0,
 			},
@@ -10853,13 +10853,13 @@ export const MODELS = {
 			reasoning: true,
 			input: ["text"],
 			cost: {
-				input: 0.08,
-				output: 0.28,
+				input: 0.09,
+				output: 0.44999999999999996,
 				cacheRead: 0,
 				cacheWrite: 0,
 			},
 			contextWindow: 40960,
-			maxTokens: 16384,
+			maxTokens: 20000,
 		} satisfies Model<"openai-completions">,
 		"qwen/qwen3-30b-a3b-instruct-2507": {
 			id: "qwen/qwen3-30b-a3b-instruct-2507",
@@ -15497,7 +15497,7 @@ export const MODELS = {
 		} satisfies Model<"openai-completions">,
 		"glm-5v-turbo": {
 			id: "glm-5v-turbo",
-			name: "glm-5v-turbo",
+			name: "GLM-5V-Turbo",
 			api: "openai-completions",
 			provider: "zai",
 			baseUrl: "https://api.z.ai/api/coding/paas/v4",